backend/requirements.txt
fastapi
psycopg2
uvicorn
pydantic
load_dotenv
os
random